Move loading logic from EITData and EITData_ to loading module
